Added buttons to each screen.
