+++ app/version.properties
@@ -1,3 +1,3 @@
-#Thu Aug 26 14:40:23 EDT 2021
-VERSION_BUILD=288
+#Fri Aug 27 12:56:35 EDT 2021
+VERSION_BUILD=310
 VERSION_NUMBER=1.0
